blog/requirements.txt
<<<<<<< HEAD
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.2
django-appconf==1.0.1
django-ckeditor==5.3.0
django-debug-toolbar==1.8
django-image-cropping==1.0.4
django-js-asset==0.1.1
django-modelcluster==3.1
django-taggit==0.22.1
django-treebeard==4.1.2
djangocms-installer==0.9.7
djangorestframework==3.6.3
easy-thumbnails==2.4.1

html5lib==0.999999999
idna==2.5
oauthlib==2.0.2
olefile==0.44
pdb==0.1
Pillow==4.0.0
psycopg2==2.7.3.1
PyJWT==1.5.2
python-decouple==3.1
python-gnupg==0.4.0
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
requests==2.18.2
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sorl-thumbnail==12.4a1
sqlparse==0.2.3
tzlocal==1.4
Unidecode==0.4.21
urllib3==1.22
virtualenv==15.1.0
webencodings==0.5.1
whitenoise==3.3.1
Willow==0.4
=======
pip install django==1.11.5

pip install django-ckeditor==5.3.0

pip install pillow==5.0





>>>>>>> 69e677a5423bdad329f9c37d7c623d661ddca6f5
